Fixed bug I introduced with coastline feature
Because I replaced the feature.geometries method with a new one
in geoaxes, it no longer called the original method from the
Feature class.  This resulted in a growing coastline feature set,
and over-drawn coastlines on the figure.  The result looked poor
and took progressively longer to calculate each time it was run.
I only noticed this when I executed in Ipython, which meant that
cartopy was loaded once, and the geometries held in memory were
augmented on every execution.