Attempt at getting nightshade to work with extended lons
This is unfinished and produces inconsistent results.  In this
implementation, the features are extended in feature artist,
which has consequences for nightshade because the reprojection
happens there instead of in Nightshade.  Also, unlike the
coastlines, nightshades involves polygons which, if unjoined,
leave gaps.  This is not trivial to solve, and may be bettter
approached by extending longitudes in the individual routines
and sending complete geometries to FeatureArtist